+++ app/caseStudies/CaseStudy.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const CaseStudy = () => {
+  return (
+    <div>
+      <h1>CaseStudy</h1>
+    </div>
+  );
+};
+
+export default CaseStudy;
