+++ app/examplePage/page.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import CaseStudy from '../caseStudies/CaseStudy';
+
+const examplePage = () => {
+  return (
+    <div>
+      <h1>
+        <CaseStudy />
+      </h1>
+    </div>
+  );
+};
+
+export default examplePage;
