--- components/CaseStudies/CaseStudies.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const page = () => {
-  return (
-    <div>
-      <h1>Case Studies</h1>
-    </div>
-  );
-};
-
-export default page;
