--- components/Contact/page.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const page = () => {
-  return (
-    <div>
-      <h1>Contact</h1>
-    </div>
-  );
-};
-
-export default page;
