--- components/Homepage/Testimonies.tsx
@@ -1,16 +1,0 @@
-const Testimonies = () => {
-  return (
-    <section className='bg-white-900 dark:bg-gray-900'>
-      <h2>
-        What
-        <div className='relative w-fit'>
-          <div className='relative z-10'>they say</div>
-          <div className='absolute bottom-0.5 z-0 h-6 w-full bg-accent' />
-        </div>
-        about me
-      </h2>
-    </section>
-  );
-};
-
-export default Testimonies;
