+++ package.json
@@ -27,6 +27,7 @@
     "postcss": "8.4.29",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-hook-form": "^7.46.2",
     "sanity": "^3.16.7",
     "styled-components": "5.2",
     "tailwindcss": "3.3.3",
